+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 - Video layers' `clipWidth` and `clipHeight` are no longer treated as invalid options.
 - Image layers' `clipWidth` and `clipHeight` are no longer set in the constructor, so if the `clipWidth` option is not supplied and the layer's `width` changes `clipWidth` uses the new `width`.
+- Video and image layers' `width` and `height` default to their `clipWidth` and `clipHeight` respectively.
 - Elliptical mask effect throwing `TypeError: path.split is not a function`.
 - In shader effects, textures whose dimensions are not powers of two accept interpolation filters.
 - No longer ignore video layers' `mediaWidth` option.